AI竞品分析_简化版/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
playwright==1.40.0
Pillow==10.2.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
